rx_frontend_gen3: Replace CORDIC and hb_dec with more optimal solutions
Specifically, we replace the CORDIC with a 1/4-rate downconverter, and
hb_dec with axi_fir.

This removes a feature from the frontend where we could apply arbitrary
IF corrections, but it reduces the resources used (and we never used the
arbitrary IF corrections). The X3x0 FPGA compat number is incremented
as the rx_frontend_gen3 is part of the X3x0 infrastructure rather than
part of an RFNoC block.
It is increased to 36 (0x24).