lib: Added converter between axi-stream and cvita.
Mostly, this is for IP Integrator's tracking of interfaces, but CHDR does not
follow convention with its bus layout. (Convention is to put first 32-bit word
in lower indices, not upper as done in CHDR IPs.) Words are swapped vs. AXI-S.

Also, the N310 implementation was updated to use the bus properly.

Signed-of-by: Alex Williams <alex.williams@ni.com>
Signed-of-by: Moritz Fischer <moritz.fischer@ettus.com>