n3xx: Move toplevel over to use the BD generated PS stuff
Move toplevel to use a BD for the processing system. This also fixes the
issue we were seeing when reloading the FPGA where the ethernet would
not work as a result of the S_AXI_GP0/1 ports getting stuck.

According to Xilinx support the way of instantiating the PS7 via RTL as
we did before is no longe supported.

Signed-off-by: Alex Williams <alex.williams.ni.com>
Signed-off-by: Moritz Fischer <moritz.fischer@ettus.com>