jesd: fix FSM reset bug and bump rev and oldest compat revision
The core is controlled by an asynchronous reset. One of the FSMs was not safely
exiting reset, causing occasional failures on one bit of the RX lanes.

Bumped the current revision and oldest compatible revision of the core to force updates
to the latest FPGA image since this is a nasty bug.