lib: Use more of DSP48 for I/Q imbalance compensation
The previous incarnation only performed the multiply in the DSP48, and
because only the M register was used, timing closure became difficult at
higher clock rates (e.g. 250 MHz for N320). This change moves the
addition inside the DSP48 and enables the P output register to shorten
the delay to exit the DSP48. This made routing easier and allowed timing
closure at 250 MHz for the frontends.