RFNoC: Align byte count to 8-byte word
The source_flow_control was counting bytes using the length in the
header while the flow_control_responder was counting the actual
words.  This caused mismatch between the byte counts in the producer
and consumer when the packet length was not word aligned, which
commonly happens when an odd number of samples is sent.  This change
aligns the packet length in the source_flow_control to the word size.

Signed-off-by: Michael West <michael.west@ettus.com>