tools: Fix for proj creation in ip_utils for tcl-based IP
The lsearch command was expecting an exact match of "_bdtcl" and the
$cmd variable content. Therefore, the commands "modify_bdtcl" and
"write_bdtcl" would result in a negative evaluation of the boolean
expression in the conditional structure; thus creating a project for
those command too (not expected).

The solution is to use the string command to look for a pattern match
of "_bdtcl" in the $cmd variable. This way, those commands that have
said suffix will skip the project creation. Whereas those that do not
have the suffix, will get a project created by the tool.