rfnoc: sim: Add test_exec header file
This includes assertion macros and time unit delcarations for use with
PkgTestExec.

Using macros for assertions allows the correct line number and file
name to be reported, but behavior is otherwise identical to calling the
PkgTestExec::assert_*() methods.

The time unit declarations are added because any file using PkgTestExec
must use the same time unit as PkgTestExec, so it makes sense to
include it in the header file.