tools: Detect error/fatal and unfinished multi-instance TBs
Errors indicated by $error() and $fatal() being called outside of our
framework were not being detected. In some cases, this caused errors to
be missed by this script.

Also, some testbenches run multiple instances of themselves, and if a
single one passed then this script would assume the testbench passed
when in fact some of the instances may have failed. By looking for
$error() and $fatal() output, we can detect when this occurs.