lib: Add internal Ethernet IPs
The ARP responder is a simple IP that responds to incoming ARP
requests. It checks for a match against its IP and sends the proper
reply if there is a match.

The eth_internal module wraps the eth_interface module and attaches
the ARP responder to the "CPU" port. This module is intended to
connect a DMA engine carrying Ethernet frames to an FPGA-internal
host. The regmap carries two sets of registers, where the DMA-side
MAC and IP addresses are the usual set, and the Bridge/Internal set
represents the FPGA-internal host. This is a PHY-less and MAC-less
Ethernet connection to RFNoC.