rfnoc: Compensate for time check delay in RX radio
This change fixes an issue where the timestamp of the first packet from
a timed RX command wouldn't match the timestamp of the command. This
change adds compensation to account for the delay required to check the
timestamp and setup the acquisition.

Note that for multiple samples per cycle, the timestamps still might
not match if the user requests a sample that ends up in the middle of a
multi-sample word from the radio.