rfnoc: Fix short packet handling
Short packets could cause multiple packets to be buffered causing a
new KV map lookup to start while the previous one was in progress. This
could cause the tuser value to be overwritten prematurely.

The fix ensures that only one lookup is in progress at a time and only
one packet goes out per lookup.