lib: Fix device_id clk-xing in rfnoc_core_kernel
The device_id input port for the rfnoc_core_kernel module is driven
in the chdr_clk domain, but it was readback by a register implemented
in the ctrl_clk domain. Thus, causing a clock crossing that the
implementation tool was trying to meet timing for.

This commit adds a short FIFO and a simple pipeline stage to safely
handle the clock-crossing.