x300: Expand DRAM address space to 1G
The address ranges configured for the AXI interconnect IP limited the
amount of accessible DRAM to two 32 MB regions. This change makes the
full 1G available to all DRAM ports.

The AXI DMA FIFO block will continue to use 32 MB by default, as
configured by the parameters on that block. But the Replay block now
has access to the full 1G by default.