image builder: Fix paths for E310 builds
This fixes a bug where the uhd_image_builder script would generate the
wrong branch for the auto-inst file for E310. This introduces a change:
The correct device type on the command line for E310 is now e31x,
following the same change that was done to the top-level directory and
auto-inst filename.

This commmand will build an image with a DDC and DUC block for the X310:

    uhd_image_builder.py -d e31x duc ddc